Remove or rename duplicate file names
This avoids problems with code coverage detection and possible double/non includes
