+++ src/ai/AIResource.h
